fix(contract): change amount data type to float

+++ Finance.Contract/Expense/CreateExpenseRequest.cs
@@ -2,6 +2,6 @@ namespace Finance.Contract.Expense;
 
 public record CreateExpenseRequest(
     string title,
-    int amount,
+    float amount,
     DateTime date
     );
